#RSS044-146 Added a retry to the sftpfilesystem connection method
1). Added some code to the Connect method so that if the connection attempt fails.  I set the number of retries to be high (25) due to the issues we have been having with datastore.  I haven't added anything to make the job sleep before retrying we can maybe add that later if this doesn't improve things.

I tried to make the number of retries be configurable in datavault,properties but for various reasons that wasn't easy.  It would have been easier to add the config to the store properties in the db but as these are stored in a hash I don't think we can change them there either.

I've added logging so that the connection attempts are in the logs

(cherry picked from commit 2f68097516970f782799424a8094bf97b2c996a2)