Make admin roles icon work on all browsers
Previously we were using the mask property to display the admin roles
icon, but that isn't supported by IE11. We switched to using a background
image which is supported by IE11, but Chrome was refusing to display it
in the correct colour. The solution is now to use the mask approach if
supported, and fall back to the background image if not